requirements.txt
# requirements.txt
fastapi>=0.115.0
httpx>=0.27.0
pydantic[email]>=2.0.0
python-dotenv>=1.0.1
supabase>=2.12.0
uvicorn[standard]>=0.30.0
